fix(ecs): finalize EntityId phase 2 with u32 generations
- migrate EntityManager.generations from Uint16Array to Uint32Array
- keep wrapping increment semantics with >>> 0
- add regression tests for >65535 generations
- add stale-ID validity test after heavy slot recycling

Validation:
- pnpm --filter @gwen/engine-core test ecs.test.ts
- pnpm --filter @gwen/engine-core test (464/464 passing)

+++ packages/@gwen/engine-core/src/core/ecs.ts
@@ -30,7 +30,7 @@ function idGeneration(id: EntityId): number {
 }
 
 export class EntityManager {
-  private generations: Uint16Array;
+  private generations: Uint32Array;
   private alive: Uint8Array; // 1 = alive, 0 = dead
   private freeList: number[] = [];
   private liveCount = 0;
@@ -38,7 +38,7 @@ export class EntityManager {
 
   constructor(maxEntities: number) {
     this.maxEntities = maxEntities;
-    this.generations = new Uint16Array(maxEntities);
+    this.generations = new Uint32Array(maxEntities);
     this.alive = new Uint8Array(maxEntities);
   }
